requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
httpx==0.27.2
anthropic==0.40.0
python-dotenv==1.0.1
